+++ src/docker/start.sh
@@ -2,6 +2,17 @@
 
 set -e
 
+### Backward compatibility
+if [ -n "$HTTP_PROXY" ]; then
+	echo "Warning: Detected proxy configuration in the old style. Please switch to passing the configuration the new way. For more information visit: https://github.com/samply/share-client/blob/master/docs/deployment/docker-deployment.md#environment-variables"
+	HTTP_PROXY_URL="$HTTP_PROXY"
+	HTTP_PROXY_USERNAME=$PROXY_USER;
+	HTTP_PROXY_PASSWORD=$PROXY_PASS;
+	HTTPS_PROXY_URL="$HTTP_PROXY"
+	HTTPS_PROXY_USERNAME=$PROXY_USER;
+	HTTPS_PROXY_PASSWORD=$PROXY_PASS;
+fi
+
 file=${CATALINA_HOME}/conf/Catalina/localhost/ROOT.xml
 sed -i "s/{postgres-host}/${POSTGRES_HOST}/"              "$file"
 sed -i "s/{postgres-port}/${POSTGRES_PORT:-5432}/"        "$file"
@@ -10,9 +21,13 @@ sed -i "s/{postgres-user}/${POSTGRES_USER}/"              "$file"
 sed -i "s/{postgres-pass}/${POSTGRES_PASS}/"              "$file"
 
 file=${CATALINA_HOME}/conf/${PROJECT}_common_config.xml
-sed -i "s~{proxy-url}~${HTTP_PROXY}~"                     "$file"
-sed -i "s/{proxy-user}/${PROXY_USER}/"                    "$file"
-sed -i "s/{proxy-pass}/${PROXY_PASS}/"                    "$file"
+sed -i "s|{http-proxy-url}|${HTTP_PROXY_URL:-}|"              "$file"
+sed -i "s|{http-proxy-username}|${HTTP_PROXY_USERNAME:-}|"    "$file"
+sed -i "s|{http-proxy-password}|${HTTP_PROXY_PASSWORD:-}|"    "$file"
+sed -i "s|{https-proxy-url}|${HTTPS_PROXY_URL:-}|"            "$file"
+sed -i "s|{https-proxy-username}|${HTTPS_PROXY_USERNAME:-}|"  "$file"
+sed -i "s|{https-proxy-password}|${HTTPS_PROXY_PASSWORD:-}|"  "$file"
+sed -i "s|{no-proxy-hosts}|${NO_PROXY_HOSTS:-}|"              "$file"
 
 file=${CATALINA_HOME}/conf/${PROJECT}_common_urls.xml
 sed -i "s#{store-url}#${STORE_URL}#"                      "$file"
@@ -22,6 +37,9 @@ sed -i "s#{directory-url}#${DIRECTORY_URL}#"              "$file"
 sed -i "s#{share-url}#${SHARE_URL}#"                      "$file"
 sed -i "s#{patientlist-url}#${PATIENTLIST_URL}#"          "$file"
 sed -i "s#{projectpseudonymisation-url}#${PROJECTPSEUDONYMISATION_URL}#"                      "$file"
+
+file=${CATALINA_HOME}/conf/secrets.properties
+sed -i "s|{id-manager-apikey}|${ID_MANAGER_APIKEY}|"      "$file"
 
 file=${CATALINA_HOME}/conf/${PROJECT}_common_operator.xml
 sed -i "s/{operator-first-name}/${OPERATOR_FIRST_NAME}/"  "$file"
@@ -35,6 +53,8 @@ sed -i "s%{mail-port}%${MAIL_PORT:-25}%"                  "$file"
 sed -i "s%{mail-protocol}%${MAIL_PROTOCOL:-smtp}%"        "$file"
 sed -i "s%{mail-from-address}%${MAIL_FROM_ADDRESS}%"      "$file"
 sed -i "s%{mail-from-name}%${MAIL_FROM_NAME}%"            "$file"
+sed -i "s%{mail-user}%${MAIL_USER}%"                      "$file"
+sed -i "s%{mail-password}%${MAIL_PASSWORD}%"              "$file"
 
 file=${CATALINA_HOME}/conf/${PROJECT}_bridgehead_info.xml
 sed -i "s#{site}#${SITE}#"                                "$file"
@@ -51,10 +71,11 @@ file=${CATALINA_HOME}/conf/log4j2.xml
 sed -i "s/{level}/${LOG_LEVEL:-info}/"                    "$file"
 
 file=${CATALINA_HOME}/conf/features.properties
-sed -i "s/{feature_BBMRI_DIRECTORY_SYNC}/${feature_BBMRI_DIRECTORY_SYNC:-false}/" "$file"
-sed -i "s/{feature_DKTK_CENTRAL_SEARCH}/${feature_DKTK_CENTRAL_SEARCH:-false}/"   "$file"
-sed -i "s/{feature_NNGM_CTS}/${feature_NNGM_CTS:-false}/"                         "$file"
-sed -i "s|{feature_NNGM_ENCRYPT_ID}|${feature_NNGM_ENCRYPT_ID:-false}|"           "$file"
+sed -i "s/{feature_bbmri_directory_sync}/${FEATURE_BBMRI_DIRECTORY_SYNC:-false}/" "$file"
+sed -i "s/{feature_dktk_central_search}/${FEATURE_DKTK_CENTRAL_SEARCH:-false}/"   "$file"
+sed -i "s/{feature_nngm_cts}/${FEATURE_NNGM_CTS:-false}/"                         "$file"
+sed -i "s|{feature_nngm_encrypt_id}|${FEATURE_NNGM_ENCRYPT_ID:-false}|"           "$file"
+sed -i "s|{feature_set_site_name}|${FEATURE_SET_SITE_NAME:-false}|"           "$file"
 
 file=${CATALINA_HOME}/conf/${PROJECT}_cts_info.xml
 sed -i "s|{nngm-magicpl-apikey}|${NNGM_MAGICPL_APIKEY}|"                    "$file"
@@ -66,8 +87,19 @@ sed -i "s|{nngm-cts-password}|${NNGM_CTS_PASSWORD}|"                        "$fi
 sed -i "s|{nngm-site-idtype}|${NNGM_SITE_IDTYPE}|"                          "$file"
 sed -i "s|{nngm-mainzelliste-apikey}|${NNGM_MAINZELLISTE_APIKEY}|"          "$file"
 sed -i "s|{nngm-mainzelliste-url}|${NNGM_MAINZELLISTE_URL}|"                "$file"
+sed -i "s|{nngm-cryptkey}|${NNGM_CRYPTKEY}|"                                "$file"
 
 export CATALINA_OPTS="${CATALINA_OPTS} -javaagent:/docker/jmx_prometheus_javaagent-0.3.1.jar=9100:/docker/jmx-exporter.yml"
+
+# SSL Certs
+if [ -d "/custom-certs" ]; then
+	echo "Found custom-certs. Starting import of certs:"
+	for file in /custom-certs/*; do
+		cp -v $file /usr/local/share/ca-certificates/$(basename $file).crt
+	done
+	update-ca-certificates || (echo -e "\nThe system has REJECTED one of the certificates:"; ls -l /custom-certs/*; echo "Make sure that ALL of the certificates are valid."; exit 1)
+	echo "Successfully imported custom-certs."
+fi
 
 # Replace start.sh with catalina.sh
 exec /usr/local/tomcat/bin/catalina.sh run