+++ src/main/java/de/samply/share/client/control/InquiryBean.java
@@ -23,7 +23,6 @@ import de.samply.share.client.model.db.tables.pojos.InquiryDetails;
 import de.samply.share.client.model.db.tables.pojos.InquiryResult;
 import de.samply.share.client.model.db.tables.pojos.InquiryResultStats;
 import de.samply.share.client.model.db.tables.pojos.RequestedEntity;
-import de.samply.share.client.quality.report.logger.PercentageLogger;
 import de.samply.share.client.rest.Connector;
 import de.samply.share.client.util.Utils;
 import de.samply.share.client.util.WebUtils;
@@ -45,6 +44,7 @@ import de.samply.share.client.util.db.InquiryUtil;
 import de.samply.share.client.util.db.UserSeenInquiryUtil;
 import de.samply.share.common.model.uiquerybuilder.QueryItem;
 import de.samply.share.common.utils.MdrIdDatatype;
+import de.samply.share.common.utils.PercentageLogger;
 import de.samply.share.common.utils.QueryTreeUtil;
 import de.samply.share.common.utils.SamplyShareUtils;
 import de.samply.share.model.bbmri.BbmriResult;
@@ -62,9 +62,13 @@ import java.io.IOException;
 import java.io.Serializable;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.Date;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
@@ -564,17 +568,19 @@ public class InquiryBean implements Serializable {
       inquiryResultsList = InquiryResultUtil
           .fetchInquiryResultsForInquiryDetailsById(latestInquiryDetails.getId());
       UserSeenInquiryUtil.setUserSeenInquiry(loginBean.getUser(), inquiry);
-      //TODO create criteriaTree with cql query
-      if (inquiryCriteria != null) {
-        if (ApplicationUtils.isLanguageQuery()) {
-          this.inquiryCriteria.add(InquiryCriteriaUtil
-              .getFirstCriteriaOriginal(latestInquiryDetails, QueryLanguageType.QL_QUERY));
-          latestOriginalCriteriaTree = populateCriteriaTree(
-              inquiryCriteria.get(0).getCriteriaOriginal());
-        } else {
-          this.inquiryCriteria = InquiryCriteriaUtil
-              .getInquiryCriteriaForInquiryDetails(latestInquiryDetails);
-        }
+      //TODO create criteriaTree with cql
+
+      if (inquiryCriteria == null) {
+        inquiryCriteria = new ArrayList<>();
+      }
+      if (ApplicationUtils.isLanguageQuery()) {
+        this.inquiryCriteria.add(InquiryCriteriaUtil
+            .getFirstCriteriaOriginal(latestInquiryDetails, QueryLanguageType.QL_QUERY));
+        latestOriginalCriteriaTree = populateCriteriaTree(
+            inquiryCriteria.get(0).getCriteriaOriginal());
+      } else {
+        this.inquiryCriteria = InquiryCriteriaUtil
+            .getInquiryCriteriaForInquiryDetails(latestInquiryDetails);
       }
 
       List<RequestedEntity> requestedEntities = InquiryUtil.getRequestedEntitiesForInquiry(inquiry);
@@ -872,7 +878,8 @@ public class InquiryBean implements Serializable {
       String filename =
           !(inquiry.getLabel().equals("")) ? inquiry.getLabel() + ".xlsx" : "Export.xlsx";
 
-      createTemporaryFile(bos, "lastExport.xlsx");
+      String lastExportFilename = generateLastExportFilename();
+      createTemporaryFile(bos, lastExportFilename);
       Faces.sendFile(bos.toByteArray(), filename, true);
 
 
@@ -880,6 +887,20 @@ public class InquiryBean implements Serializable {
       logger.error("Exception caught while trying to export data", e);
     }
   }
+
+  private String generateLastExportFilename() {
+    String timestampForFilename = getTimestampForFilename(new Date());
+
+    return "export-" + timestampForFilename + ".xlsx";
+  }
+
+  private String getTimestampForFilename(Date timestamp) {
+
+    DateFormat simpleDateFormat = new SimpleDateFormat("yyyyMMdd-HHmm", Locale.ENGLISH);
+    return simpleDateFormat.format(timestamp);
+
+  }
+
 
   private void createTemporaryFile(ByteArrayOutputStream byteArrayOutputStream, String filename) {