+++ src/main/java/de/samply/share/client/quality/report/file/excel/instances/statistics/dataelementstats/DataElementStatsExcelRowContext.java
@@ -2,10 +2,10 @@ package de.samply.share.client.quality.report.file.excel.instances.statistics.da
 
 import de.samply.share.client.quality.report.file.excel.row.context.ExcelRowContextImpl;
 import de.samply.share.client.quality.report.file.excel.row.elements.ExcelRowElements;
-import de.samply.share.client.quality.report.logger.PercentageLogger;
 import de.samply.share.client.quality.report.results.QualityResults;
 import de.samply.share.client.quality.report.results.statistics.QualityResultsStatistics;
 import de.samply.share.common.utils.MdrIdDatatype;
+import de.samply.share.common.utils.PercentageLogger;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;