+++ src/main/java/de/samply/share/client/quality/report/file/excel/sheet/ExcelSheetFactoryImpl.java
@@ -4,12 +4,13 @@ import de.samply.share.client.quality.report.file.excel.row.context.ExcelRowCont
 import de.samply.share.client.quality.report.file.excel.row.elements.ExcelRowElements;
 import de.samply.share.client.quality.report.file.excel.row.factory.ExcelRowFactory;
 import de.samply.share.client.quality.report.file.excel.row.factory.ExcelRowFactoryException;
-import de.samply.share.client.quality.report.logger.PercentageLogger;
+import de.samply.share.common.utils.PercentageLogger;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.poi.ss.SpreadsheetVersion;
-import org.apache.poi.xssf.usermodel.XSSFSheet;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.apache.poi.xssf.streaming.SXSSFSheet;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
+
 
 public class ExcelSheetFactoryImpl implements ExcelSheetFactory {
 
@@ -22,11 +23,12 @@ public class ExcelSheetFactoryImpl implements ExcelSheetFactory {
   }
 
   @Override
-  public XSSFWorkbook addSheet(XSSFWorkbook workbook, String sheetTitle,
+  public SXSSFWorkbook addSheet(SXSSFWorkbook workbook, String sheetTitle,
       ExcelRowContext excelRowContext) throws ExcelSheetFactoryException {
 
-    XSSFSheet sheet = workbook.createSheet(sheetTitle);
+    SXSSFSheet sheet = workbook.createSheet(sheetTitle);
     sheet = addRowTitles(sheet, excelRowContext);
+    sheet.trackAllColumnsForAutoSizing();
 
     int maxNumberOfRows = SpreadsheetVersion.EXCEL2007.getMaxRows();
 
@@ -49,7 +51,7 @@ public class ExcelSheetFactoryImpl implements ExcelSheetFactory {
 
   }
 
-  private XSSFSheet addRowTitles(XSSFSheet sheet, ExcelRowContext excelRowContext)
+  private SXSSFSheet addRowTitles(SXSSFSheet sheet, ExcelRowContext excelRowContext)
       throws ExcelSheetFactoryException {
 
     try {
@@ -62,7 +64,7 @@ public class ExcelSheetFactoryImpl implements ExcelSheetFactory {
 
   }
 
-  private XSSFSheet addRow(XSSFSheet sheet, ExcelRowElements excelRowElements)
+  private SXSSFSheet addRow(SXSSFSheet sheet, ExcelRowElements excelRowElements)
       throws ExcelSheetFactoryException {
 
     try {