+++ src/main/java/de/samply/share/client/quality/report/file/excel/sheet/wrapper/ExcelSheetFactoryWrapper.java
@@ -3,8 +3,8 @@ package de.samply.share.client.quality.report.file.excel.sheet.wrapper;
 import de.samply.share.client.quality.report.file.excel.row.context.ExcelRowContext;
 import de.samply.share.client.quality.report.file.excel.sheet.ExcelSheetFactory;
 import de.samply.share.client.quality.report.file.excel.sheet.ExcelSheetFactoryException;
-import org.apache.poi.xssf.usermodel.XSSFSheet;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.apache.poi.xssf.streaming.SXSSFSheet;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
 
 public abstract class ExcelSheetFactoryWrapper implements ExcelSheetFactory {
 
@@ -14,16 +14,16 @@ public abstract class ExcelSheetFactoryWrapper implements ExcelSheetFactory {
     this.excelSheetFactory = excelSheetFactory;
   }
 
-  protected abstract XSSFSheet addFunctionalityToSheet(XSSFSheet sheet);
+  protected abstract SXSSFSheet addFunctionalityToSheet(SXSSFSheet sheet);
 
-  private XSSFWorkbook addFunctionalityToSheet(XSSFWorkbook workbook, String sheetTitle) {
-    XSSFSheet sheet = workbook.getSheet(sheetTitle);
+  private SXSSFWorkbook addFunctionalityToSheet(SXSSFWorkbook workbook, String sheetTitle) {
+    SXSSFSheet sheet = workbook.getSheet(sheetTitle);
     addFunctionalityToSheet(sheet);
     return workbook;
   }
 
   @Override
-  public XSSFWorkbook addSheet(XSSFWorkbook workbook, String sheetTitle,
+  public SXSSFWorkbook addSheet(SXSSFWorkbook workbook, String sheetTitle,
       ExcelRowContext excelRowContext) throws ExcelSheetFactoryException {
 
     workbook = excelSheetFactory.addSheet(workbook, sheetTitle, excelRowContext);