+++ src/main/java/de/samply/share/client/quality/report/file/excel/sheet/wrapper/HighlightNotMappedInOrangeExcelSheetFactory002.java
@@ -3,9 +3,10 @@ package de.samply.share.client.quality.report.file.excel.sheet.wrapper;
 import de.samply.share.client.quality.report.file.excel.cell.element.MatchElement;
 import de.samply.share.client.quality.report.file.excel.row.elements.ExcelRowElements002;
 import de.samply.share.client.quality.report.file.excel.sheet.ExcelSheetFactory;
-import org.apache.poi.hssf.util.CellReference;
+import org.apache.poi.ss.usermodel.FontFormatting;
 import org.apache.poi.ss.usermodel.IndexedColors;
-import org.apache.poi.xssf.usermodel.XSSFFontFormatting;
+import org.apache.poi.ss.util.CellReference;
+
 
 public class HighlightNotMappedInOrangeExcelSheetFactory002 extends
     ExcelSheetWithHighLightFactory {
@@ -43,7 +44,7 @@ public class HighlightNotMappedInOrangeExcelSheetFactory002 extends
   }
 
   @Override
-  protected XSSFFontFormatting setHighlightFontFormatting(XSSFFontFormatting fontFormatting) {
+  protected FontFormatting setHighlightFontFormatting(FontFormatting fontFormatting) {
 
     fontFormatting.setFontColorIndex(IndexedColors.ORANGE.getIndex());
     return fontFormatting;