+++ src/main/java/de/samply/share/client/util/connector/exception/MainzellisteConnectorException.java
@@ -1,0 +1,8 @@
+package de.samply.share.client.util.connector.exception;
+
+public class MainzellisteConnectorException extends Exception {
+
+  public MainzellisteConnectorException(Throwable cause) {
+    super(cause);
+  }
+}